+++ source/StdInc.h
@@ -173,7 +173,7 @@
 #include "VehicleScanner.h"
 #include "LoadMonitor.h"
 #include "PedStuckChecker.h"
-#include "DecisionMakerTypes.h"
+#include "DecisionMakers/DecisionMakerTypes.h"
 #include "InformGroupEventQueue.h"
 #include "InformFriendsEventQueue.h"
 #include "Events/GroupEventHandler.h"
@@ -261,9 +261,9 @@
 #include "CutsceneMgr.h"
 #include "Darkel.h"
 #include "Date.h"
-#include "Decision.h"
-#include "DecisionMaker.h"
-#include "DecisionSimple.h"
+#include "DecisionMakers/Decision.h"
+#include "DecisionMakers/DecisionMaker.h"
+#include "DecisionMakers/DecisionSimple.h"
 #include "Directory.h"
 #include "Draw.h"
 #include "EntityScanner.h"
