+++ source/game_sa/DecisionMakers/Decision.cpp
@@ -2,8 +2,16 @@
 
 #include "Decision.h"
 
-void CDecision::InjectHooks() {
+void CDecision::InjectHooks() {    
+    RH_ScopedClass(CDecision);
+    RH_ScopedCategory("DecisionMakers");
 
+    RH_ScopedInstall(SetDefault, 0x600530, { .reversed = false });
+    RH_ScopedInstall(Set, 0x600570);
+    //RH_ScopedInstall(Add, 0x600600, { .reversed = false });
+    RH_ScopedInstall(From, 0x6006B0);
+    //RH_ScopedInstall(HasResponse, 0x600710, { .reversed = false });
+    //RH_ScopedInstall(MakeDecision, 0x6040D0, { .reversed = false });
 }
 
 // 0x6040C0
