+++ source/game_sa/DecisionMakers/DecisionSimple.cpp
