+++ source/game_sa/DecisionMakers/DecisionSimple.h
