+++ source/game_sa/PedGroupIntelligence.h
@@ -9,7 +9,7 @@
 #include <concepts>
 #include <PedGroup.h>
 #include "PedTaskPair.h"
-#include "DecisionMakerTypes.h"
+#include "DecisionMakers/DecisionMakerTypes.h"
 #include <TaskManager.h>
 #include "Tasks/TaskTypes/TaskSimpleNone.h"
 #include "Tasks/Allocators/PedGroup/PedGroupDefaultTaskAllocatorType.h"
